refactor: component themes
