+++ packages/styled-ui-theme/src/foundations/borderStyles.js
@@ -1,0 +1,3 @@
+const borderStyles = {};
+
+export default borderStyles;
