+++ packages/styled-ui-theme/src/foundations/borderWidths.js
@@ -1,0 +1,7 @@
+const borderWidths = {
+  none: 0,
+  1: '1px',
+  2: '2px',
+};
+
+export default borderWidths;
