+++ packages/styled-ui-theme/src/foundations/borders.js
@@ -1,0 +1,7 @@
+const borders = {
+  none: 0,
+  1: '1px solid',
+  2: '2px solid',
+};
+
+export default borders;
