+++ packages/styled-ui-theme/src/foundations/fontSizes.js
@@ -1,0 +1,12 @@
+const fontSizes = {
+  xs: '.75rem',
+  sm: '.875rem',
+  md: '1rem',
+  lg: '1.125rem',
+  xl: '1.25rem',
+  '2xl': '1.5rem',
+  '3xl': '1.75rem',
+  '4xl': '2rem',
+};
+
+export default fontSizes;
