+++ packages/styled-ui-theme/src/foundations/fontWeights.js
@@ -1,0 +1,13 @@
+const fontWeights = {
+  thin: 100,
+  extralight: 200,
+  light: 300,
+  normal: 400,
+  medium: 500,
+  semibold: 600,
+  bold: 700,
+  extrabold: 800,
+  black: 900,
+};
+
+export default fontWeights;
