+++ packages/styled-ui-theme/src/foundations/fonts.js
@@ -1,0 +1,19 @@
+/**
+ * A generic font family should be the last item in the list of font family names:
+ * - serif
+ * - sans-serif
+ * - monospace
+ * - cursive
+ * - fantasy
+ * - system-ui
+ * - math
+ * - emoji
+ * - fangsong
+ */
+const fonts = {
+  heading: '"Segoe UI", -apple-system, BlinkMacSystemFont, "Helvetica Neue", Helvetica, Arial, sans-serif',
+  base: '"Segoe UI", -apple-system, BlinkMacSystemFont, "Helvetica Neue", Helvetica, Arial, sans-serif',
+  mono: '"Segoe UI Mono", "SFMono-Medium", "SF Mono", Menlo, Consolas, Courier, monospace',
+};
+
+export default fonts;
