+++ packages/styled-ui-theme/src/foundations/letterSpacings.js
@@ -1,0 +1,3 @@
+const letterSpacings = {};
+
+export default letterSpacings;
