+++ packages/styled-ui-theme/src/foundations/lineHeights.js
@@ -1,0 +1,14 @@
+const lineHeights = {
+  normal: 'normal',
+  base: '1.5',
+  xs: '1.125rem',
+  sm: '1.25rem',
+  md: '1.375rem',
+  lg: '1.5rem',
+  xl: '1.75rem',
+  '2xl': '2rem',
+  '3xl': '2.25rem',
+  '4xl': '2.5rem',
+};
+
+export default lineHeights;
