+++ packages/styled-ui-theme/src/foundations/radii.js
@@ -1,0 +1,9 @@
+const radii = {
+  circle: '50%',
+  none: 0,
+  sm: '.1875rem',
+  md: '.375rem',
+  lg: '.75rem',
+};
+
+export default radii;
