+++ packages/styled-ui-theme/src/foundations/sizes.js
@@ -1,0 +1,7 @@
+import space from './space';
+
+const sizes = {
+  ...space,
+};
+
+export default sizes;
