+++ packages/styled-ui-theme/src/foundations/space.js
@@ -1,21 +1,19 @@
-import base from './base';
-
-const spaceUnit = 'px';
+const spaceUnit = 'rem';
 const spaceDefinition = {
   quarter: {
     list: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
     notation: 'q',
-    value: 1,
+    value: 0.0625,
   },
   half: {
     list: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
     notation: 'h',
-    value: 2,
+    value: 0.125,
   },
   whole: {
     list: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 24, 32, 40, 48, 56, 64],
     notation: 'x',
-    value: 4,
+    value: 0.25,
   },
 };
 
@@ -42,49 +40,4 @@ const space = (() => {
   return accumulatedSpace;
 })();
 
-const fontSizes = {
-  xs: '12px',
-  sm: '14px',
-  md: '16px',
-  lg: '18px',
-  xl: '20px',
-  '2xl': '24px',
-  '3xl': '28px',
-  '4xl': '32px',
-};
-
-const lineHeights = {
-  normal: 'normal',
-  base: '1.5',
-  xs: '18px',
-  sm: '20px',
-  md: '22px',
-  lg: '24px',
-  xl: '28px',
-  '2xl': '32px',
-  '3xl': '36px',
-  '4xl': '40px',
-};
-
-const sizes = {
-  ...space,
-};
-
-const radii = {
-  circle: '50%',
-  none: 0,
-  sm: '3px',
-  md: '6px',
-  lg: '12px',
-};
-
-const theme = {
-  ...base,
-  space,
-  fontSizes,
-  lineHeights,
-  sizes,
-  radii,
-};
-
-export default theme;
+export default space;
