+++ packages/styled-ui-theme/src/foundations/zIndices.js
@@ -1,0 +1,16 @@
+const zIndices = {
+  hide: -1,
+  auto: 'auto',
+  base: 0,
+  dropdown: 1000,
+  sticky: 1100,
+  fixed: 1200,
+  overlay: 1300,
+  drawer: 1400,
+  modal: 1500,
+  popover: 1600,
+  toast: 1700,
+  tooltip: 1800,
+};
+
+export default zIndices;
