+++ packages/styled-ui-theme/src/index.js
@@ -1,9 +1,7 @@
 import base from './base';
-import fixed from './fixed';
 import getColorPalette from './color-palette';
 
 export {
   base,
-  fixed,
   getColorPalette,
 };
